+++ org.eclipse.ice.dev.annotations/src/test/java/org/eclipse/ice/tests/dev/annotations/processors/FieldBuilderTest.java
@@ -1,4 +1,4 @@
-package org.eclipse.ice.tests.dev.annotations;
+package org.eclipse.ice.tests.dev.annotations.processors;
 
 import static org.junit.jupiter.api.Assertions.*;
